Extract render request logic to separate module and following refactoring
